Modern_Application_Development_2- Library Management System/Code/requirements.txt
amqp==5.1.1
aniso8601==9.0.1
async-timeout==4.0.2
autopep8==2.0.1
billiard==3.6.4.0
blinker==1.5
Brotli==1.0.9
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
chardet==5.2.0
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
cors==1.0.1
cryptography==38.0.4
cssselect2==0.7.0
cvxpy==1.2.2
dnspython==2.3.0
ecos==2.0.12
email-validator==1.3.1
filelock==3.8.2
Flask==2.2.2
Flask-Cors==3.0.10
Flask-Excel==0.0.7
Flask-JWT==0.2.0
Flask-Login==0.6.2
Flask-Principal==0.4.0
flask-redis==0.4.0
Flask-RESTful==0.3.9
Flask-Security-Too==5.1.2
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.1.1
fonttools==4.38.0
future==0.18.2
gevent==22.10.2
greenlet==2.0.1
html5lib==1.1
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.2.4
lml==0.1.0
mapping==0.1.6
MarkupSafe==2.1.1
numpy==1.24.0
osqp==0.6.2.post8
pandas==1.5.2
passlib==1.7.4
Pillow==9.3.0
pip==22.0.2
prompt-toolkit==3.0.36
pycodestyle==2.10.0
pycparser==2.21
pydyf==0.5.0
pyexcel==0.7.0
pyexcel-io==0.6.6
pyexcel-webio==0.1.4
PyJWT==2.6.0
pyphen==0.13.2
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.7
qdldl==0.1.5.post2
redis==4.4.0
requests==2.28.1
requests-file==1.5.1
scipy==1.9.3
scs==3.2.2
setuptools==59.6.0
six==1.16.0
SQLAlchemy==1.4.45
texttable==1.7.0
tinycss2==1.2.1
tldextract==3.4.0
tomli==2.0.1
typing-extensions==4.5.0
tzdata==2023.3
urllib3==1.26.13
vine==5.0.0
wcwidth==0.2.5
weasyprint==57.2
webencodings==0.5.1
Werkzeug==2.2.2
WTForms==3.0.1
zipp==3.15.0
zope.event==4.6
zope.interface==5.5.2
zopfli==0.2.2
